Move common files to common
