+++ src/common/formatTools.ts
