+++ src/middleware/verify-jwt.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from "express";
 import { decodeJwtToken } from "../services/auth/auth-lib";
 import jsonwebtoken from "jsonwebtoken";
 import { StatusCode } from "status-code-enum";
-import Config from "../config";
+import Config from "../common/config";
 
 // TODO: Remove all usages of these
