+++ src/services/admission/admission-router.test.ts
@@ -2,9 +2,9 @@ import { beforeEach, describe, expect, it, jest } from "@jest/globals";
 import Models from "../../database/models";
 import { DecisionStatus, DecisionResponse, AdmissionDecision } from "../../database/admission-db";
 import { RegistrationFormat } from "../registration/registration-formats";
-import { RegistrationTemplates } from "./../../config";
+import { RegistrationTemplates } from "../../common/config";
 import { Gender, Degree, Race, HackInterest, HackOutreach } from "../registration/registration-models";
-import { getAsStaff, getAsUser, putAsStaff, putAsUser, getAsAttendee, putAsApplicant, TESTER } from "../../testTools";
+import { getAsStaff, getAsUser, putAsStaff, putAsUser, getAsAttendee, putAsApplicant, TESTER } from "../../common/testTools";
 import { StatusCode } from "status-code-enum";
 import type * as MailLib from "../../services/mail/mail-lib";
 import type { AxiosResponse } from "axios";
