+++ src/services/registration/registration-formats.ts
@@ -1,5 +1,5 @@
 import { Degree, Gender, HackInterest, HackOutreach, Race } from "./registration-models";
-import { isString, isBoolean, isArrayOfType, isNumber, isEnumOfType } from "../../formatTools";
+import { isString, isBoolean, isArrayOfType, isNumber, isEnumOfType } from "../../common/formatTools";
 
 export interface RegistrationFormat {
     userId: string;
