+++ src/services/registration/registration-router.test.ts
@@ -2,8 +2,8 @@ import { beforeEach, describe, expect, it, jest } from "@jest/globals";
 import { StatusCode } from "status-code-enum";
 import Models from "../../database/models";
 import { RegistrationApplication } from "../../database/registration-db";
-import { RegistrationTemplates } from "./../../config";
-import { TESTER, getAsUser, getAsAdmin, postAsUser } from "../../testTools";
+import { RegistrationTemplates } from "../../common/config";
+import { TESTER, getAsUser, getAsAdmin, postAsUser } from "../../common/testTools";
 import { RegistrationFormat } from "./registration-formats";
 import { Degree, Gender, HackInterest, HackOutreach, Race } from "./registration-models";
 import type * as MailLib from "../../services/mail/mail-lib";
