+++ src/services/staff/staff-router.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it } from "@jest/globals";
-import { AUTH_ROLE_TO_ROLES, putAsAttendee, putAsStaff } from "../../testTools";
+import { AUTH_ROLE_TO_ROLES, putAsAttendee, putAsStaff } from "../../common/testTools";
 import { generateJwtToken } from "../auth/auth-lib";
 
 import { EventAttendance } from "../../database/event-db";
