+++ src/services/user/user-router.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, afterEach, describe, expect, it } from "@jest/globals";
-import { AUTH_ROLE_TO_ROLES, TESTER, get, getAsAdmin, getAsAttendee, getAsStaff, putAsAttendee } from "../../testTools";
+import { AUTH_ROLE_TO_ROLES, TESTER, get, getAsAdmin, getAsAttendee, getAsStaff, putAsAttendee } from "../../common/testTools";
 
 import { AttendeeFollowing, AttendeeProfile } from "../../database/attendee-db";
 import { EventFollowers, EventAttendance, Event } from "../../database/event-db";
 import { StatusCode } from "status-code-enum";
-import Config from "../../config";
+import Config from "../../common/config";
 import { AuthInfo } from "../../database/auth-db";
 import Models from "../../database/models";
 import { UserInfo } from "../../database/user-db";
